# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgenticAudit",
    "description": "Log, classify and audit AI agent actions for GDPR/AI Act/SOC2 compliance",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/compliance/ai-act/status": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "AI Act compliance status",
        "description": "Get overall AI Act compliance status for the organization.",
        "operationId": "compliance_status_v1_compliance_ai_act_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Compliance Status V1 Compliance Ai Act Status Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/compliance/ai-act/report": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Download AI Act compliance report PDF",
        "description": "Generate and download an AI Act compliance report.",
        "operationId": "compliance_report_pdf_v1_compliance_ai_act_report_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/compliance/ai-act/fria/{system_id}/pdf": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Download FRIA PDF for a system",
        "description": "Generate a pre-filled FRIA PDF for a specific AI system.",
        "operationId": "fria_pdf_v1_compliance_ai_act_fria__system_id__pdf_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}